mq队列问题修复

+++ Lumina_api/operations/base_view.py
@@ -8,18 +8,23 @@ class BaseView(APIView):
     create_log = False
     models = None
     serializer = None
-    get_filter = None
+    allowed_methods = ['get', 'post', 'patch', 'delete']
+
+    def get_queryset(self, request, *args, **kwargs):
+        queryset = self.models.objects.all()
+        return queryset
 
     def get(self, request, row_id=None):
-        if self.get_filter:
-            queryset = self.models.objects.filter(**{f'{self.get_filter}': row_id}).all()
-            data = get_data(queryset, True, request, self, self.serializer)
-        else:
-            data = get_data(self.models, False, request, self, self.serializer)
+        if 'get' not in self.allowed_methods:
+            return JsonResponse(return_response(status=False, info='请求方法错误！'))
+        queryset = self.get_queryset(request, row_id)
+        data = get_data(queryset, True, request, self, self.serializer)
         response = return_response(data=data)
         return JsonResponse(response)
 
     def post(self, request):
+        if 'post' not in self.allowed_methods:
+            return JsonResponse(return_response(status=False, info='请求方法错误！'))
         ser = self.serializer(data=request.data)
         if ser.is_valid():
             ser.save()
@@ -31,6 +36,8 @@ class BaseView(APIView):
         return JsonResponse(response)
 
     def patch(self, request, row_id):
+        if 'patch' not in self.allowed_methods:
+            return JsonResponse(return_response(status=False, info='请求方法错误！'))
         queryset = self.models.objects.get(id=row_id)
         ser = self.serializer(instance=queryset, data=request.data)
         if ser.is_valid():
@@ -43,6 +50,8 @@ class BaseView(APIView):
         return JsonResponse(response)
 
     def delete(self, request, row_id):
+        if 'delete' not in self.allowed_methods:
+            return JsonResponse(return_response(status=False, info='请求方法错误！'))
         try:
             data = self.models.objects.filter(id=row_id).delete()
             if self.create_log:
